Include the inherents in `processing_bundles`
With some internal discussion, Nazar provided a nice clue that we
inject the inherents in the very end when processing the bundles,
in that way #233 can be eliminated and it's simpler and intuitive as
well.

Fixes #233
